Move TransferForm from components to pages
